+++ frameworks/preact/src/types/field.ts
@@ -8,11 +8,7 @@ import type {
   ValidPath,
 } from '@formisch/core/preact';
 import type { ReadonlySignal } from '@preact/signals';
-import type {
-  FocusEventHandler,
-  GenericEventHandler,
-  InputEventHandler,
-} from 'preact';
+import type { GenericEventHandler, InputEventHandler } from 'preact';
 import type * as v from 'valibot';
 
 /**
@@ -34,7 +30,7 @@ export interface FieldElementProps {
   /**
    * The focus event handler of the field element.
    */
-  readonly onFocus: FocusEventHandler<FieldElement>;
+  readonly onFocus: () => void;
   /**
    * The input event handler of the field element.
    */
@@ -46,7 +42,7 @@ export interface FieldElementProps {
   /**
    * The blur event handler of the field element.
    */
-  readonly onBlur: FocusEventHandler<FieldElement>;
+  readonly onBlur: () => void;
 }
 
 /**